refactor: extract Remote and HtmlTagDescriptor types into src/types/
Amp-Thread-ID: https://ampcode.com/threads/T-019d3464-cde5-708b-9128-0bbe1c9a30cb
