+++ packages/lib/src/types/html.ts
@@ -1,0 +1,9 @@
+export interface HtmlTagDescriptor {
+  tag: string;
+  attrs?: Record<string, string | boolean | undefined>;
+  children?: string | HtmlTagDescriptor[];
+  /**
+   * default: 'head-prepend'
+   */
+  injectTo?: 'head' | 'body' | 'head-prepend' | 'body-prepend';
+}
