+++ packages/lib/src/types/remote.ts
@@ -1,0 +1,7 @@
+import type { RemotesConfig } from 'types';
+
+export type Remote = {
+  id: string;
+  regexp: RegExp;
+  config: RemotesConfig;
+};
